+++ sdk/python/feast/infra/online_stores/duckdb_online_store/duckdb.py
@@ -7,7 +7,7 @@ import duckdb
 from feast import Entity
 from feast.feature_view import FeatureView
 from feast.infra.key_encoding_utils import serialize_entity_key
-from feast.infra.online_stores.helpers import _table_id, _to_naive_utc
+from feast.infra.online_stores.helpers import _table_id, _to_naive_utc, _process_rows
 from feast.infra.online_stores.online_store import OnlineStore
 from feast.infra.online_stores.vector_store import VectorStoreConfig
 from feast.protos.feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
@@ -111,7 +111,7 @@ class DuckDBOnlineStore(OnlineStore):
         """
         try:
             rows = conn.execute(query, keys).fetchall()
-            return self._process_rows(keys, rows)
+            return _process_rows(keys, rows)
         except Exception as e:
             raise RuntimeError(f"Failed to read from DuckDB: {e}")
         finally:
@@ -244,33 +244,3 @@ class DuckDBOnlineStore(OnlineStore):
             raise RuntimeError(f"Failed to create index in DuckDB: {e}")
         finally:
             self._close_conn()
-
-    @staticmethod
-    def _process_rows(
-        keys: List[bytes], rows: List[Tuple]
-    ) -> List[Tuple[Optional[datetime], Optional[Dict[str, ValueProto]]]]:
-        """Transform the retrieved rows in the desired output.
-
-        PostgreSQL may return rows in an unpredictable order. Therefore, `values_dict`
-        is created to quickly look up the correct row using the keys, since these are
-        actually in the correct order.
-        """
-        values_dict = defaultdict(list)
-        for row in rows if rows is not None else []:
-            values_dict[
-                row[0] if isinstance(row[0], bytes) else row[0].tobytes()
-            ].append(row[1:])
-
-        result: List[Tuple[Optional[datetime], Optional[Dict[str, ValueProto]]]] = []
-        for key in keys:
-            if key in values_dict:
-                value = values_dict[key]
-                res = {}
-                for feature_name, value_bin, event_ts in value:
-                    val = ValueProto()
-                    val.ParseFromString(bytes(value_bin))
-                    res[feature_name] = val
-                result.append((event_ts, res))
-            else:
-                result.append((None, None))
-        return result
